Particle system (#338)
* base structure

* ayo v1 of particles

* small bug fixing

* shape goog

* particless

* Ayo particles

* update docs

* update changelog

* make tests pass

* son of a bitch

* linting

* nice demo

* density

* scale actually working

* raise notimplemented

* clean a few things

* typehint int

* fix things a do demo

* demo file

* Delete function rework (#342)

* delete stuff gooder

* clean up animation

* clean spritesheet

* fix docs

* wowie

* Update CHANGELOG.md

* revert p_demo

* Update CHANGELOG.md

* clean surface.py

* free old surface when overriding

* optimize raster and give from_file args

* fix docs and demos

* b ruh

* remove deprecated methods


* Update particles_demo.py

* lord help me with this

* random mode takes things into account

* move updating of particles

* better asteroid particles

* goog movement for demo

* move the draw

* restructure particlesystem

* world_space and local_space

* Update asteroids.py

* fix rotation

* fix rotation again

* accelerations

* particle system gen

* reignore demo cuz lazy

* bugfixes and demo improvement

* more small cleanups

* wipe feature_test

* running

* slowly but surely fix pyright errors

* big chungus fixes

* bye bye

* Demo + changelog

* particle documentation better

* more better docs

* Update rubato/utils/rb_time.py


* Make custom_comps good

+++ demo/physics_demo.py
@@ -66,8 +66,8 @@ for _ in range(num_obj // 2):
                 ),
             ],
             pos=(
-                randint(rb.Display.res.x / 20, 19 * rb.Display.res.x / 20),
-                randint(rb.Display.res.y / 20, 19 * rb.Display.res.y / 20),
+                randint(int(rb.Display.res.x / 20), int(19 * rb.Display.res.x / 20)),
+                randint(int(rb.Display.res.y / 20), int(19 * rb.Display.res.y / 20)),
             ),
         )
     )
@@ -85,8 +85,8 @@ for _ in range(num_obj // 2):
                 ),
             ],
             pos=(
-                randint(rb.Display.res.x / 20, 19 * rb.Display.res.x / 20),
-                randint(rb.Display.res.y / 20, 19 * rb.Display.res.y / 20),
+                randint(int(rb.Display.res.x / 20), int(19 * rb.Display.res.x / 20)),
+                randint(int(rb.Display.res.y / 20), int(19 * rb.Display.res.y / 20)),
             ),
         )
     )
